# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for eCRF Application",
        "title": "eCRF API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/protocols/add": {
            "post": {
                "description": "creating a new protocol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocol"
                ],
                "summary": "Protocol Add",
                "parameters": [
                    {
                        "description": "addprotocol",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.AddProtocol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.AddProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/protocols/delete": {
            "delete": {
                "description": "Deleting a  protocol  by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocol"
                ],
                "summary": "Protocol Delete",
                "parameters": [
                    {
                        "description": "deleteprotocol",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteProtocol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DeleteProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/protocols/save": {
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "saving of updating prtocol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocol"
                ],
                "summary": "Protocol Save",
                "parameters": [
                    {
                        "description": "saveprotocol",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SaveProtocol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SaveProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/protocols/{filter}/{center}": {
            "get": {
                "description": "Getting all protocols from users center id and filter params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocol"
                ],
                "summary": "Protocols Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter",
                        "name": "filter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Center",
                        "name": "center",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.GetProtocols"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
        "/api/subject/add": {
            "post": {
                "description": "creating a new protocol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subject"
                ],
                "summary": "Subject Add",
                "parameters": [
                    {
                        "description": "addsubject",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.AddSubject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.AddProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
        "/api/subject/{protocol_id}": {
            "get": {
                "description": "Getting all subjects by protocol id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subject"
                ],
                "summary": "Subjects  Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "protocol_id",
                        "name": "protocol_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.GetSubject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/subject/{protocol_id}/{subject_num}": {
            "get": {
                "description": "Getting all subjects by protocol id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subject"
                ],
                "summary": "Subjects  Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "protocol_id",
                        "name": "protocol_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "subject_num",
                        "name": "subject_num",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.GetSubject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.AddProtocol": {
            "type": "object",
            "properties": {
                "centerID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "requests.AddSubject": {
            "type": "object",
            "properties": {
                "center_id": {
                    "type": "integer"
                },
                "initials": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "protocol_id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.DeleteProtocol": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.SaveProtocol": {
            "type": "object",
            "properties": {
                "centerId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "responses.AddProtocol": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.DeleteProtocol": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.GetProtocols": {
            "type": "object",
            "properties": {
                "center_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "protocol_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "responses.GetSubject": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "center_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "initials": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "protocol_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "responses.SaveProtocol": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Cookie",
            "in": "Cookie"
        }
    }
}